refactor(desktop): split chat window range controller
