+++ apps/desktop/src/components/features/agents/agent-chat/use-agent-chat-window.ts
@@ -1,18 +1,11 @@
 import type { RefCallback, RefObject } from "react";
-import { useCallback, useEffect, useLayoutEffect, useRef, useState } from "react";
+import { useLayoutEffect } from "react";
 import type { AgentChatWindowRow } from "./agent-chat-thread-windowing";
-import { CHAT_OVERSCAN, CHAT_SHIFT_SIZE, CHAT_WINDOW_SIZE } from "./agent-chat-thread-windowing";
-import {
-  CHAT_AUTO_SCROLL_ANIMATION_DURATION_MS,
-  CHAT_MAX_RENDERED_ROWS,
-  clampWindowRange,
-  createBottomAnchoredWindow,
-  EMPTY_WINDOW,
-  type WindowRange,
-} from "./agent-chat-window-shared";
+import { clampWindowRange, createBottomAnchoredWindow } from "./agent-chat-window-shared";
 import { useAgentChatResizeSync } from "./use-agent-chat-resize-sync";
 import { useAgentChatScrollController } from "./use-agent-chat-scroll-controller";
 import { useAgentChatSentinelObservers } from "./use-agent-chat-sentinel-observers";
+import { useAgentChatWindowRangeController } from "./use-agent-chat-window-range-controller";
 
 type UseAgentChatWindowInput = {
   rows: AgentChatWindowRow[];
@@ -44,10 +37,6 @@ export function useAgentChatWindow({
 }: UseAgentChatWindowInput): UseAgentChatWindowResult {
   const rowCount = rows.length;
   const initialWindow = createBottomAnchoredWindow(rowCount);
-  const [windowRange, setWindowRange] = useState<WindowRange>(() => initialWindow);
-  const prevSessionIdRef = useRef<string | null>(null);
-  const prevIsSessionViewLoadingRef = useRef(isSessionViewLoading);
-  const prevRowCountRef = useRef(rowCount);
   const {
     isNearBottom,
     isNearTop,
@@ -68,80 +57,6 @@ export function useAgentChatWindow({
     initialWindow,
   });
 
-  const applyBottomAnchoredWindow = useCallback(() => {
-    const nextWindow = createBottomAnchoredWindow(rowCount);
-    setWindowRange(nextWindow);
-    setBottomAnchoredState(nextWindow.start);
-  }, [rowCount, setBottomAnchoredState]);
-
-  useEffect(() => {
-    if (prevSessionIdRef.current === activeSessionId) {
-      return;
-    }
-
-    prevSessionIdRef.current = activeSessionId;
-    applyBottomAnchoredWindow();
-    requestWindowScroll({
-      target: "bottom",
-      behavior: "auto",
-      suppressSentinels: true,
-    });
-  }, [activeSessionId, applyBottomAnchoredWindow, requestWindowScroll]);
-
-  useEffect(() => {
-    const finishedLoading = prevIsSessionViewLoadingRef.current && !isSessionViewLoading;
-    prevIsSessionViewLoadingRef.current = isSessionViewLoading;
-    if (!finishedLoading) {
-      return;
-    }
-
-    applyBottomAnchoredWindow();
-    requestWindowScroll({
-      target: "bottom",
-      behavior: "auto",
-      suppressSentinels: true,
-    });
-  }, [applyBottomAnchoredWindow, isSessionViewLoading, requestWindowScroll]);
-
-  useEffect(() => {
-    const previousRowCount = prevRowCountRef.current;
-    prevRowCountRef.current = rowCount;
-
-    if (rowCount === 0) {
-      setWindowRange(EMPTY_WINDOW);
-      setBottomAnchoredState(0);
-      return;
-    }
-
-    if (rowCount < previousRowCount) {
-      setWindowRange((current) => clampWindowRange(current, rowCount));
-      return;
-    }
-
-    if (rowCount === previousRowCount || !isPinnedToBottomRef.current) {
-      return;
-    }
-
-    applyBottomAnchoredWindow();
-    if (hasPendingScrollRequest()) {
-      return;
-    }
-
-    requestWindowScroll({
-      target: "bottom",
-      behavior: "auto",
-      suppressSentinels: false,
-      animationDurationMs: CHAT_AUTO_SCROLL_ANIMATION_DURATION_MS,
-    });
-  }, [
-    applyBottomAnchoredWindow,
-    hasPendingScrollRequest,
-    isPinnedToBottomRef,
-    requestWindowScroll,
-    rowCount,
-    setBottomAnchoredState,
-  ]);
-
   useLayoutEffect(() => {
     applyPendingScrollRequest();
   });
@@ -151,86 +66,22 @@ export function useAgentChatWindow({
     messagesContentRef,
     syncBottomIfPinned,
   });
-
-  const shiftWindowUp = useCallback(() => {
-    if (isUpdatingRef.current || rowCount === 0 || suppressSentinelsRef.current) {
-      return;
-    }
-
-    setWindowRange((current) => {
-      if (current.start <= 0) {
-        return current;
-      }
-
-      const nextStart = Math.max(0, current.start - CHAT_SHIFT_SIZE);
-      const nextRange = clampWindowRange(
-        {
-          start: nextStart,
-          end: Math.min(rowCount - 1, nextStart + CHAT_MAX_RENDERED_ROWS - 1),
-        },
-        rowCount,
-      );
-      if (nextRange.start === current.start && nextRange.end === current.end) {
-        return current;
-      }
-
-      const container = messagesContainerRef.current;
-      if (container) {
-        prevScrollHeightRef.current = container.scrollHeight;
-        shouldCompensateScrollRef.current = true;
-      }
-      isUpdatingRef.current = true;
-      if (typeof window !== "undefined") {
-        window.requestAnimationFrame(() => {
-          isUpdatingRef.current = false;
-        });
-      } else {
-        isUpdatingRef.current = false;
-      }
-      return nextRange;
+  const { windowRange, scrollToBottom, scrollToTop, shiftWindowUp, shiftWindowDown } =
+    useAgentChatWindowRangeController({
+      rowCount,
+      activeSessionId,
+      isSessionViewLoading,
+      messagesContainerRef,
+      isPinnedToBottomRef,
+      suppressSentinelsRef,
+      prevScrollHeightRef,
+      shouldCompensateScrollRef,
+      isUpdatingRef,
+      hasPendingScrollRequest,
+      requestWindowScroll,
+      setBottomAnchoredState,
+      setTopAnchoredState,
     });
-  }, [
-    isUpdatingRef,
-    messagesContainerRef,
-    prevScrollHeightRef,
-    rowCount,
-    shouldCompensateScrollRef,
-    suppressSentinelsRef,
-  ]);
-
-  const shiftWindowDown = useCallback(() => {
-    if (isUpdatingRef.current || rowCount === 0 || suppressSentinelsRef.current) {
-      return;
-    }
-
-    setWindowRange((current) => {
-      if (current.end >= rowCount - 1) {
-        return current;
-      }
-
-      const nextEnd = Math.min(rowCount - 1, current.end + CHAT_SHIFT_SIZE);
-      const nextRange = clampWindowRange(
-        {
-          start: Math.max(0, nextEnd - CHAT_MAX_RENDERED_ROWS + 1),
-          end: nextEnd,
-        },
-        rowCount,
-      );
-      if (nextRange.start === current.start && nextRange.end === current.end) {
-        return current;
-      }
-
-      isUpdatingRef.current = true;
-      if (typeof window !== "undefined") {
-        window.requestAnimationFrame(() => {
-          isUpdatingRef.current = false;
-        });
-      } else {
-        isUpdatingRef.current = false;
-      }
-      return nextRange;
-    });
-  }, [isUpdatingRef, rowCount, suppressSentinelsRef]);
 
   const { topSentinelRef, bottomSentinelRef } = useAgentChatSentinelObservers({
     messagesContainerRef,
@@ -244,37 +95,11 @@ export function useAgentChatWindow({
 
   const effectiveIsNearTop = isNearTop && windowRange.start === 0;
 
-  const scrollToBottom = useCallback(() => {
-    applyBottomAnchoredWindow();
-    requestWindowScroll({
-      target: "bottom",
-      behavior: "auto",
-      suppressSentinels: true,
-    });
-  }, [applyBottomAnchoredWindow, requestWindowScroll]);
-
-  const scrollToTop = useCallback(() => {
-    const nextRange = clampWindowRange(
-      {
-        start: 0,
-        end: Math.min(rowCount - 1, CHAT_WINDOW_SIZE + CHAT_OVERSCAN - 1),
-      },
-      rowCount,
-    );
-    setWindowRange(nextRange);
-    setTopAnchoredState();
-    requestWindowScroll({
-      target: "top",
-      behavior: "auto",
-      suppressSentinels: true,
-    });
-  }, [requestWindowScroll, rowCount, setTopAnchoredState]);
-
   const effectiveWindow = clampWindowRange(windowRange, rowCount);
-  const windowedRows =
-    effectiveWindow.end >= effectiveWindow.start
-      ? rows.slice(effectiveWindow.start, effectiveWindow.end + 1)
-      : [];
+  const hasVisibleRows = effectiveWindow.end >= effectiveWindow.start;
+  const windowedRows = hasVisibleRows
+    ? rows.slice(effectiveWindow.start, effectiveWindow.end + 1)
+    : [];
 
   return {
     windowedRows,
